[Create] 코맨트 컴포넌트 분리
수정/삭제 기능 분리 해서 작업 합니다
